test(activemodel): redistribute Validations / Types / Validators-extended (PR 26 cont.)
Continues the misc.test.ts split from #821. Pure test movement, no
rename/reword.

- Validations (130-566) — whole block (incl. all 11 nested per-validator
  describes + loose-it tail) -> validations.test.ts as a sibling
  describe. Keeps the parent context so test:compare sees
  Validations > presence/absence/etc. unchanged.
- Types (570-685) — split each nested describe into its existing
  type/<name>.test.ts: DateType, DateTimeType, DecimalType, TypeRegistry.
- Validators (extended) (694-838) — split into the matching
  validations/<name>-validation.test.ts files: format-with-without,
  numericality-comparison-operators, inclusion-allowNil, exclusion-allowNil,
  acceptance-skips-nil; "custom messages" -> validations.test.ts.
- Drop the now-stale Phase 1100/1150 / Errors.on() / Validators-untested
  separator headers in misc.test.ts. Drop unused `Types` import.

misc.test.ts: 2,064 -> 1,353 lines. Test count unchanged (1,535 AM).
