+++ packages/activemodel/src/misc.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { Model, Types } from "./index.js";
+import { Model } from "./index.js";
 
 describe("ActiveModel", () => {
   // =========================================================================
@@ -121,718 +121,6 @@ describe("ActiveModel", () => {
       expect(admin.readAttribute("role")).toBe("admin");
       expect(Admin.attributeNames()).toContain("name");
       expect(Admin.attributeNames()).toContain("role");
-    });
-  });
-
-  // =========================================================================
-  // Phase 1100/1150 — Validations
-  // =========================================================================
-  describe("Validations", () => {
-    // -- Presence --
-    describe("presence", () => {
-      class Article extends Model {
-        static {
-          this.attribute("title", "string");
-          this.validates("title", { presence: true });
-        }
-      }
-
-      it("rejects null", () => {
-        const a = new Article();
-        expect(a.isValid()).toBe(false);
-        expect(a.errors.get("title")).toContain("can't be blank");
-      });
-
-      it("rejects empty string", () => {
-        const a = new Article({ title: "" });
-        expect(a.isValid()).toBe(false);
-      });
-
-      it("rejects whitespace-only string", () => {
-        const a = new Article({ title: "   " });
-        expect(a.isValid()).toBe(false);
-      });
-
-      it("accepts a real value", () => {
-        const a = new Article({ title: "Hello" });
-        expect(a.isValid()).toBe(true);
-      });
-    });
-
-    // -- Absence --
-    describe("absence", () => {
-      class Blank extends Model {
-        static {
-          this.attribute("name", "string");
-          this.validates("name", { absence: true });
-        }
-      }
-
-      it("accepts null", () => {
-        expect(new Blank().isValid()).toBe(true);
-      });
-
-      it("accepts empty string", () => {
-        expect(new Blank({ name: "" }).isValid()).toBe(true);
-      });
-
-      it("rejects a value", () => {
-        const b = new Blank({ name: "dean" });
-        expect(b.isValid()).toBe(false);
-        expect(b.errors.get("name")).toContain("must be blank");
-      });
-    });
-
-    // -- Length --
-    describe("length", () => {
-      class WithLength extends Model {
-        static {
-          this.attribute("name", "string");
-          this.validates("name", {
-            length: { minimum: 3, maximum: 10 },
-          });
-        }
-      }
-
-      it("validates length of using minimum", () => {
-        const w = new WithLength({ name: "ab" });
-        expect(w.isValid()).toBe(false);
-        expect(w.errors.get("name")[0]).toMatch(/is too short/);
-      });
-
-      it("validates length of using maximum", () => {
-        const w = new WithLength({ name: "abcdefghijk" });
-        expect(w.isValid()).toBe(false);
-        expect(w.errors.get("name")[0]).toMatch(/is too long/);
-      });
-
-      it("validates length of using within", () => {
-        expect(new WithLength({ name: "dean" }).isValid()).toBe(true);
-      });
-
-      it("validates length of using is", () => {
-        class Exact extends Model {
-          static {
-            this.attribute("code", "string");
-            this.validates("code", { length: { is: 4 } });
-          }
-        }
-        expect(new Exact({ code: "1234" }).isValid()).toBe(true);
-        expect(new Exact({ code: "123" }).isValid()).toBe(false);
-        expect(new Exact({ code: "12345" }).isValid()).toBe(false);
-      });
-
-      it("validates with in (range)", () => {
-        class WithRange extends Model {
-          static {
-            this.attribute("name", "string");
-            this.validates("name", { length: { in: [2, 5] } });
-          }
-        }
-        expect(new WithRange({ name: "a" }).isValid()).toBe(false);
-        expect(new WithRange({ name: "ab" }).isValid()).toBe(true);
-        expect(new WithRange({ name: "abcde" }).isValid()).toBe(true);
-        expect(new WithRange({ name: "abcdef" }).isValid()).toBe(false);
-      });
-
-      it("skips null values (null has no length)", () => {
-        expect(new WithLength({}).isValid()).toBe(true);
-      });
-    });
-
-    // -- Numericality --
-    describe("numericality", () => {
-      class Numeric extends Model {
-        static {
-          this.attribute("value", "string"); // string to test cast behavior
-          this.validates("value", { numericality: true });
-        }
-      }
-
-      it("default validates numericality of", () => {
-        expect(new Numeric({ value: "42" }).isValid()).toBe(true);
-        expect(new Numeric({ value: "3.14" }).isValid()).toBe(true);
-      });
-
-      it("rejects non-numeric strings", () => {
-        const n = new Numeric({ value: "not a number" });
-        expect(n.isValid()).toBe(false);
-        expect(n.errors.get("value")).toContain("is not a number");
-      });
-
-      it("validates numericality of with nil allowed", () => {
-        expect(new Numeric({}).isValid()).toBe(true);
-      });
-
-      it("validates numericality of with integer only", () => {
-        class IntOnly extends Model {
-          static {
-            this.attribute("count", "string");
-            this.validates("count", { numericality: { onlyInteger: true } });
-          }
-        }
-        expect(new IntOnly({ count: "5" }).isValid()).toBe(true);
-        const f = new IntOnly({ count: "5.5" });
-        expect(f.isValid()).toBe(false);
-        expect(f.errors.get("count")).toContain("must be an integer");
-      });
-
-      it("validates numericality with greater than", () => {
-        class GT extends Model {
-          static {
-            this.attribute("age", "integer");
-            this.validates("age", { numericality: { greaterThan: 0 } });
-          }
-        }
-        expect(new GT({ age: 1 }).isValid()).toBe(true);
-        expect(new GT({ age: 0 }).isValid()).toBe(false);
-      });
-
-      it("validates numericality with less than", () => {
-        class LT extends Model {
-          static {
-            this.attribute("rating", "integer");
-            this.validates("rating", { numericality: { lessThan: 10 } });
-          }
-        }
-        expect(new LT({ rating: 9 }).isValid()).toBe(true);
-        expect(new LT({ rating: 10 }).isValid()).toBe(false);
-      });
-
-      it("validates numericality with odd", () => {
-        class Odd extends Model {
-          static {
-            this.attribute("n", "integer");
-            this.validates("n", { numericality: { odd: true } });
-          }
-        }
-        expect(new Odd({ n: 3 }).isValid()).toBe(true);
-        expect(new Odd({ n: 4 }).isValid()).toBe(false);
-      });
-
-      it("validates numericality with even", () => {
-        class Even extends Model {
-          static {
-            this.attribute("n", "integer");
-            this.validates("n", { numericality: { even: true } });
-          }
-        }
-        expect(new Even({ n: 4 }).isValid()).toBe(true);
-        expect(new Even({ n: 3 }).isValid()).toBe(false);
-      });
-    });
-
-    // -- Inclusion / Exclusion --
-    describe("inclusion", () => {
-      class Status extends Model {
-        static {
-          this.attribute("status", "string");
-          this.validates("status", { inclusion: { in: ["draft", "published"] } });
-        }
-      }
-
-      it("validates inclusion of", () => {
-        expect(new Status({ status: "draft" }).isValid()).toBe(true);
-      });
-
-      it("rejects non-included values", () => {
-        const s = new Status({ status: "invalid" });
-        expect(s.isValid()).toBe(false);
-        expect(s.errors.get("status")).toContain("is not included in the list");
-      });
-    });
-
-    describe("exclusion", () => {
-      class NoAdmin extends Model {
-        static {
-          this.attribute("role", "string");
-          this.validates("role", { exclusion: { in: ["admin", "root"] } });
-        }
-      }
-
-      it("accepts non-excluded values", () => {
-        expect(new NoAdmin({ role: "user" }).isValid()).toBe(true);
-      });
-
-      it("validates exclusion of", () => {
-        const n = new NoAdmin({ role: "admin" });
-        expect(n.isValid()).toBe(false);
-        expect(n.errors.get("role")).toContain("is reserved");
-      });
-    });
-
-    // -- Format --
-    describe("format", () => {
-      class Email extends Model {
-        static {
-          this.attribute("email", "string");
-          this.validates("email", { format: { with: /^[^@]+@[^@]+$/ } });
-        }
-      }
-
-      it("validate format", () => {
-        expect(new Email({ email: "dean@example.com" }).isValid()).toBe(true);
-      });
-
-      it("rejects non-matching format", () => {
-        const e = new Email({ email: "not-an-email" });
-        expect(e.isValid()).toBe(false);
-        expect(e.errors.get("email")).toContain("is invalid");
-      });
-
-      it("skips null", () => {
-        expect(new Email({}).isValid()).toBe(true);
-      });
-    });
-
-    // -- Acceptance --
-    describe("acceptance", () => {
-      // Rails' equivalent uses a virtual attribute; boolean here so `"1"` /
-      // `true` round-trip through cast as `true` and match the default accept
-      // list `["1", true]`, whereas a string-typed attr would cast `true` to
-      // "t" (per type/immutable_string.rb) and rightly fail.
-      class Terms extends Model {
-        static {
-          this.attribute("accepted", "boolean");
-          this.validates("accepted", { acceptance: true });
-        }
-      }
-
-      it("terms of service agreement", () => {
-        expect(new Terms({ accepted: "1" }).isValid()).toBe(true);
-        expect(new Terms({ accepted: true }).isValid()).toBe(true);
-      });
-
-      it("terms of service agreement no acceptance", () => {
-        expect(new Terms({ accepted: "0" }).isValid()).toBe(false);
-        expect(new Terms({ accepted: false }).isValid()).toBe(false);
-      });
-
-      it("terms of service agreement with accept value", () => {
-        class Custom extends Model {
-          static {
-            this.attribute("agreed", "string");
-            this.validates("agreed", {
-              acceptance: { accept: ["I agree", "yes"] },
-            });
-          }
-        }
-        expect(new Custom({ agreed: "I agree" }).isValid()).toBe(true);
-        expect(new Custom({ agreed: "no" }).isValid()).toBe(false);
-      });
-    });
-
-    // -- Confirmation --
-    describe("confirmation", () => {
-      class WithConfirm extends Model {
-        static {
-          this.attribute("password", "string");
-          this.validates("password", { confirmation: true });
-        }
-      }
-
-      it("passes when no confirmation field set", () => {
-        expect(new WithConfirm({ password: "secret" }).isValid()).toBe(true);
-      });
-
-      it("title confirmation", () => {
-        expect(
-          new WithConfirm({
-            password: "secret",
-            passwordConfirmation: "secret",
-          }).isValid(),
-        ).toBe(true);
-      });
-
-      it("no title confirmation", () => {
-        const w = new WithConfirm({
-          password: "secret",
-          passwordConfirmation: "wrong",
-        });
-        expect(w.isValid()).toBe(false);
-        expect(w.errors.get("passwordConfirmation")).toContain("doesn't match Password");
-      });
-    });
-
-    // -- Conditional validation --
-    describe("conditional", () => {
-      it("if validation using block false", () => {
-        class Cond extends Model {
-          static {
-            this.attribute("name", "string");
-            this.attribute("requireName", "boolean", { default: false });
-            this.validates("name", {
-              presence: {
-                if: (r: any) => r.readAttribute("requireName") === true,
-              },
-            });
-          }
-        }
-        expect(new Cond({ requireName: false }).isValid()).toBe(true);
-        expect(new Cond({ requireName: true }).isValid()).toBe(false);
-      });
-
-      it("unless validation using block true", () => {
-        class Unless extends Model {
-          static {
-            this.attribute("name", "string");
-            this.attribute("optional", "boolean", { default: false });
-            this.validates("name", {
-              presence: {
-                unless: (r: any) => r.readAttribute("optional") === true,
-              },
-            });
-          }
-        }
-        expect(new Unless({ optional: true }).isValid()).toBe(true);
-        expect(new Unless({ optional: false }).isValid()).toBe(false);
-      });
-    });
-
-    // -- Custom validate --
-    describe("custom validate", () => {
-      it("function validator", () => {
-        class Custom extends Model {
-          static {
-            this.attribute("value", "integer");
-            this.validate(function (record: any) {
-              const val = record.readAttribute("value");
-              if (val !== null && (val as number) % 2 !== 0) {
-                record.errors.add("value", "even", { message: "must be even" });
-              }
-            });
-          }
-        }
-        expect(new Custom({ value: 4 }).isValid()).toBe(true);
-        const c = new Custom({ value: 3 });
-        expect(c.isValid()).toBe(false);
-        expect(c.errors.get("value")).toContain("must be even");
-      });
-    });
-
-    // -- isInvalid --
-    it("invalid should be the opposite of valid", () => {
-      class Required extends Model {
-        static {
-          this.attribute("name", "string");
-          this.validates("name", { presence: true });
-        }
-      }
-      expect(new Required().isInvalid()).toBe(true);
-      expect(new Required({ name: "dean" }).isInvalid()).toBe(false);
-    });
-
-    // -- fullMessages --
-    it("fullMessages prefixes attribute name", () => {
-      class FM extends Model {
-        static {
-          this.attribute("title", "string");
-          this.validates("title", { presence: true });
-        }
-      }
-      const f = new FM();
-      f.isValid();
-      expect(f.errors.fullMessages).toContain("Title can't be blank");
-    });
-
-    it("fullMessages for :base has no prefix", () => {
-      class Base extends Model {
-        static {
-          this.validate((record: any) => {
-            record.errors.add("base", "invalid", { message: "is broken" });
-          });
-        }
-      }
-      const b = new Base();
-      b.isValid();
-      expect(b.errors.fullMessages).toContain("is broken");
-    });
-
-    // -- errors.clear between validations --
-    it("errors are cleared between isValid calls", () => {
-      class Clearable extends Model {
-        static {
-          this.attribute("name", "string");
-          this.validates("name", { presence: true });
-        }
-      }
-      const c = new Clearable();
-      c.isValid();
-      expect(c.errors.count).toBeGreaterThan(0);
-      c.writeAttribute("name", "dean");
-      c.isValid();
-      expect(c.errors.count).toBe(0);
-    });
-  });
-
-  // Errors / Callbacks / Dirty Tracking suites moved to dedicated files.
-
-  describe("Types", () => {
-    describe("DateType", () => {
-      const type = new Types.DateType();
-
-      it("has name 'date'", () => {
-        expect(type.name).toBe("date");
-      });
-
-      it("casts Date to Date", () => {
-        const d = new Date("2024-01-15");
-        expect(type.cast(d)).toBe(d);
-      });
-
-      it("casts null to null", () => {
-        expect(type.cast(null)).toBe(null);
-      });
-
-      it("casts undefined to null", () => {
-        expect(type.cast(undefined)).toBe(null);
-      });
-
-      it("casts invalid string to null", () => {
-        expect(type.cast("not-a-date")).toBe(null);
-      });
-
-      it("deserialize delegates to cast", () => {
-        const result = type.deserialize("2024-01-15");
-        expect(result).toBeInstanceOf(Date);
-      });
-
-      it("serialize delegates to cast", () => {
-        const result = type.serialize("2024-01-15");
-        expect(result).toBeInstanceOf(Date);
-      });
-    });
-
-    describe("DateTimeType", () => {
-      const type = new Types.DateTimeType();
-
-      it("has name 'datetime'", () => {
-        expect(type.name).toBe("datetime");
-      });
-
-      it("casts null to null", () => {
-        expect(type.cast(null)).toBe(null);
-      });
-
-      it("casts Date to Date", () => {
-        const d = new Date();
-        expect(type.cast(d)).toBe(d);
-      });
-    });
-
-    describe("DecimalType", () => {
-      const type = new Types.DecimalType();
-
-      it("has name 'decimal'", () => {
-        expect(type.name).toBe("decimal");
-      });
-
-      it("type cast decimal", () => {
-        expect(type.cast(42.5)).toBe("42.5");
-      });
-
-      it("casts string number to string", () => {
-        expect(type.cast("3.14")).toBe("3.14");
-      });
-
-      it("casts integer to string", () => {
-        expect(type.cast(100)).toBe("100");
-      });
-
-      it("casts null to null", () => {
-        expect(type.cast(null)).toBe(null);
-      });
-
-      it("type cast decimal from invalid string", () => {
-        // Mirrors Rails decimal_test.rb — "" nils out; leading-numeric
-        // prefix is kept; no-numeric-prefix returns BigDecimal(0).
-        expect(type.cast("")).toBe(null);
-        expect(type.cast("1ignore")).toBe("1");
-        expect(type.cast("bad1")).toBe("0");
-        expect(type.cast("bad")).toBe("0");
-      });
-    });
-
-    describe("TypeRegistry", () => {
-      it("looks up built-in types", () => {
-        const str = Types.typeRegistry.lookup("string");
-        expect(str).toBeInstanceOf(Types.StringType);
-      });
-
-      it("looks up integer type", () => {
-        const int = Types.typeRegistry.lookup("integer");
-        expect(int).toBeInstanceOf(Types.IntegerType);
-      });
-
-      it("looks up all built-in types", () => {
-        expect(Types.typeRegistry.lookup("float")).toBeInstanceOf(Types.FloatType);
-        expect(Types.typeRegistry.lookup("boolean")).toBeInstanceOf(Types.BooleanType);
-        expect(Types.typeRegistry.lookup("date")).toBeInstanceOf(Types.DateType);
-        expect(Types.typeRegistry.lookup("datetime")).toBeInstanceOf(Types.DateTimeType);
-        expect(Types.typeRegistry.lookup("decimal")).toBeInstanceOf(Types.DecimalType);
-      });
-
-      it("a reasonable error is given when no type is found", () => {
-        expect(() => Types.typeRegistry.lookup("imaginary")).toThrow("Unknown type: imaginary");
-      });
-
-      it("a class can be registered for a symbol", () => {
-        Types.typeRegistry.register("custom", () => new Types.StringType());
-        const t = Types.typeRegistry.lookup("custom");
-        expect(t).toBeInstanceOf(Types.StringType);
-      });
-    });
-  });
-
-  // =========================================================================
-  // Errors.on() alias
-  // =========================================================================
-
-  // =========================================================================
-  // Validators — untested options
-  // =========================================================================
-  describe("Validators (extended)", () => {
-    describe("format with 'without' option", () => {
-      class NoNumbers extends Model {
-        static {
-          this.attribute("name", "string");
-          this.validates("name", { format: { without: /\d/ } });
-        }
-      }
-
-      it("accepts values not matching 'without'", () => {
-        expect(new NoNumbers({ name: "dean" }).isValid()).toBe(true);
-      });
-
-      it("rejects values matching 'without'", () => {
-        const n = new NoNumbers({ name: "dean123" });
-        expect(n.isValid()).toBe(false);
-        expect(n.errors.get("name")).toContain("is invalid");
-      });
-    });
-
-    describe("numericality comparison operators", () => {
-      it("validates numericality with greater than or equal", () => {
-        class GTE extends Model {
-          static {
-            this.attribute("age", "integer");
-            this.validates("age", { numericality: { greaterThanOrEqualTo: 18 } });
-          }
-        }
-        expect(new GTE({ age: 18 }).isValid()).toBe(true);
-        expect(new GTE({ age: 17 }).isValid()).toBe(false);
-      });
-
-      it("validates numericality with less than or equal to", () => {
-        class LTE extends Model {
-          static {
-            this.attribute("rating", "integer");
-            this.validates("rating", { numericality: { lessThanOrEqualTo: 5 } });
-          }
-        }
-        expect(new LTE({ rating: 5 }).isValid()).toBe(true);
-        expect(new LTE({ rating: 6 }).isValid()).toBe(false);
-      });
-
-      it("validates numericality with equal to", () => {
-        class EQ extends Model {
-          static {
-            this.attribute("answer", "integer");
-            this.validates("answer", { numericality: { equalTo: 42 } });
-          }
-        }
-        expect(new EQ({ answer: 42 }).isValid()).toBe(true);
-        expect(new EQ({ answer: 41 }).isValid()).toBe(false);
-      });
-
-      it("validates numericality with other than", () => {
-        class OT extends Model {
-          static {
-            this.attribute("count", "integer");
-            this.validates("count", { numericality: { otherThan: 0 } });
-          }
-        }
-        expect(new OT({ count: 1 }).isValid()).toBe(true);
-        expect(new OT({ count: 0 }).isValid()).toBe(false);
-      });
-    });
-
-    describe("inclusion allowNil", () => {
-      it("validates inclusion of with allow nil", () => {
-        class WithNil extends Model {
-          static {
-            this.attribute("status", "string");
-            this.validates("status", { inclusion: { in: ["a", "b"] } });
-          }
-        }
-        expect(new WithNil({}).isValid()).toBe(true);
-      });
-
-      it("validates nil when allowNil is false", () => {
-        class NoNil extends Model {
-          static {
-            this.attribute("status", "string");
-            this.validates("status", { inclusion: { in: ["a", "b"], allowNil: false } });
-          }
-        }
-        expect(new NoNil({}).isValid()).toBe(false);
-      });
-    });
-
-    describe("exclusion allowNil", () => {
-      it("skips nil by default", () => {
-        class WithNil extends Model {
-          static {
-            this.attribute("role", "string");
-            this.validates("role", { exclusion: { in: ["admin"] } });
-          }
-        }
-        expect(new WithNil({}).isValid()).toBe(true);
-      });
-    });
-
-    describe("acceptance skips nil", () => {
-      it("skips nil by default", () => {
-        class Terms extends Model {
-          static {
-            this.attribute("terms", "string");
-            this.validates("terms", { acceptance: true });
-          }
-        }
-        expect(new Terms({}).isValid()).toBe(true);
-      });
-    });
-
-    describe("custom messages", () => {
-      it("presence with custom message", () => {
-        class Custom extends Model {
-          static {
-            this.attribute("name", "string");
-            this.validates("name", { presence: { message: "is required" } });
-          }
-        }
-        const c = new Custom();
-        c.isValid();
-        expect(c.errors.get("name")).toContain("is required");
-      });
-
-      it("length with custom tooShort and tooLong", () => {
-        class Custom extends Model {
-          static {
-            this.attribute("name", "string");
-            this.validates("name", {
-              length: { minimum: 3, maximum: 5, tooShort: "too few!", tooLong: "too many!" },
-            });
-          }
-        }
-        const short = new Custom({ name: "ab" });
-        short.isValid();
-        expect(short.errors.get("name")).toContain("too few!");
-
-        const long = new Custom({ name: "abcdef" });
-        long.isValid();
-        expect(long.errors.get("name")).toContain("too many!");
-      });
     });
   });
